refactor: export loadExtend
changes

- loadApplication > loadApplicationExtend
- loadContext > loadContextExtend
- loadRequest > loadRequestExtend
- loadResponse > loadResponseExtend
- loadHelper > loadHelperExtend
