+++ lib/base_loader.js
@@ -5,7 +5,7 @@ const path = require('path');
 const assert = require('assert');
 const isFunction = require('is-type-of').function;
 const interopRequire = require('interop-require');
-const debug = require('debug')('egg:loader');
+const debug = require('debug')('egg-loader');
 const Loader = require('./loader');
 const Emitter = require('events').EventEmitter;
 
@@ -288,7 +288,7 @@ class EggLoader {
 const loaders = [
   require('./plugin_loader'),
   require('./config_loader'),
-  require('./extend_loader'),
+  require('./loader/extend'),
   require('./proxy_loader'),
   require('./service_loader'),
   require('./middleware_loader'),
