+++ lib/loader/extend.js
@@ -3,10 +3,8 @@
 const fs = require('fs');
 const path = require('path');
 const interopRequire = require('interop-require');
-const utils = require('./utils');
-const debug = require('debug')('egg:extend:loader');
-
-const loadExtend = Symbol('loadExtend');
+const debug = require('debug')('egg-loader:extend');
+const utils = require('../utils');
 
 module.exports = {
 
@@ -14,61 +12,61 @@ module.exports = {
    * 扩展 Agent.prototype 的属性
    *
    * 可加载路径查看 {@link EggLoader#loadDirs}
-   * @method EggLoader#loadAgent
+   * @method EggLoader#loadAgentExtend
    */
-  loadAgent() {
-    this[loadExtend]('agent', this.app);
+  loadAgentExtend() {
+    this.loadExtend('agent', this.app);
   },
 
   /**
    * 扩展 Application.prototype 的属性
    *
    * 可加载路径查看 {@link EggLoader#loadDirs}
-   * @method EggLoader#loadApplication
+   * @method EggLoader#loadApplicationExtend
    */
-  loadApplication() {
-    this[loadExtend]('application', this.app);
+  loadApplicationExtend() {
+    this.loadExtend('application', this.app);
   },
 
   /**
    * 扩展 Request.prototype 的属性
    *
    * 可加载路径查看 {@link EggLoader#loadDirs}
-   * @method EggLoader#loadRequest
+   * @method EggLoader#loadRequestExtend
    */
-  loadRequest() {
-    this[loadExtend]('request', this.app.request);
+  loadRequestExtend() {
+    this.loadExtend('request', this.app.request);
   },
 
   /**
    * 扩展 Response.prototype 的属性
    *
    * 可加载路径查看 {@link EggLoader#loadDirs}
-   * @method EggLoader#loadResponse
+   * @method EggLoader#loadResponseExtend
    */
-  loadResponse() {
-    this[loadExtend]('response', this.app.response);
+  loadResponseExtend() {
+    this.loadExtend('response', this.app.response);
   },
 
   /**
    * 扩展 Context.prototype 的属性
    *
    * 可加载路径查看 {@link EggLoader#loadDirs}
-   * @method EggLoader#loadContext
+   * @method EggLoader#loadContextExtend
    */
-  loadContext() {
-    this[loadExtend]('context', this.app.context);
+  loadContextExtend() {
+    this.loadExtend('context', this.app.context);
   },
 
   /**
    * 扩展 app.Helper.prototype 的属性
    *
    * 可加载路径查看 {@link EggLoader#loadDirs}
-   * @method EggLoader#loadHelper
+   * @method EggLoader#loadHelperExtend
    */
-  loadHelper() {
+  loadHelperExtend() {
     if (this.app && this.app.Helper) {
-      this[loadExtend]('helper', this.app.Helper.prototype);
+      this.loadExtend('helper', this.app.Helper.prototype);
     }
   },
 
@@ -80,7 +78,7 @@ module.exports = {
    * @param {Object} proto - 最终将属性合并到 proto 上
    * @private
    */
-  [loadExtend](name, proto) {
+  loadExtend(name, proto) {
     // 获取需要加载的文件
     const filepaths = this.loadDirs()
       .map(dir => {
